Legacy/graph_prac/requirements.txt
pip
python-igraph
